Create CollapsedStoryStateActions component
